backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.30.0
pydantic==2.7.1
PyJWT==2.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
